# src/main/resources/影院订票系统.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "影院订票系统",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/sessions": {
      "get": {
        "summary": "查询电影场次",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "movieId",
            "in": "query",
            "description": "电影ID  ",
            "required": false,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "场次日期(格式yyyy-MM-dd)  ",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "start_time": {
                        "type": "string"
                      },
                      "end_time": {
                        "type": "string"
                      },
                      "price": {
                        "type": "number"
                      },
                      "hall": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "type"
                        ]
                      }
                    }
                  }
                },
                "example": {
                  "id": "Long, 场次ID",
                  "start_time": "DateTime, 开场时间",
                  "end_time": "DateTime, 结束时间",
                  "price": "Double, 基础价格",
                  "hall": {
                    "id": "Long, 影厅ID",
                    "name": "String, 影厅名称",
                    "type": "String, 影厅类型"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/sessions/{sessionId}/seats": {
      "get": {
        "summary": "获取场次座位图",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "场次ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session_id": {
                      "type": "integer"
                    },
                    "seats": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "row_no": {
                            "type": "string"
                          },
                          "col_no": {
                            "type": "integer"
                          },
                          "type": {
                            "type": "string"
                          },
                          "price_factor": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "required": [
                    "session_id",
                    "seats"
                  ]
                },
                "example": {
                  "session_id": "Long",
                  "seats": [
                    {
                      "id": "Long, 座位ID",
                      "row_no": "String, 行号",
                      "col_no": "Integer, 列号",
                      "type": "String, 座位类型",
                      "price_factor": "Double, 价格系数",
                      "status": "String, 状态(AVAILABLE/LOCKED)"
                    }
                  ]
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/sessions": {
      "post": {
        "summary": "创建放映场次",
        "deprecated": false,
        "description": "为指定电影创建新场次",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "场次创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/sessions/{sessionId}": {
      "put": {
        "summary": "调整场次信息",
        "deprecated": false,
        "description": "修改场次时间、影厅或票价",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "取消场次",
        "deprecated": false,
        "description": "取消指定场次并通知相关用户",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "取消成功",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/admin/sessions/{sessionId}/seats": {
      "get": {
        "summary": "监控场次售票情况",
        "deprecated": false,
        "description": "获取场次座位销售状态",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "座位状态图",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeatMapResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "type": "object",
        "required": [
          "movieId",
          "hallId",
          "startTime",
          "price"
        ],
        "properties": {
          "movieId": {
            "type": "integer",
            "description": "电影ID"
          },
          "hallId": {
            "type": "integer",
            "description": "影厅ID"
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "开场时间"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "基础票价"
          }
        }
      },
...
      "UpdateSessionRequest": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "format": "date-time",
            "description": "新开场时间"
          },
          "hallId": {
            "type": "integer",
            "description": "新影厅ID"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "新票价"
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "场次ID"
          },
          "movieId": {
            "type": "integer",
            "description": "电影ID"
          },
          "hallId": {
            "type": "integer",
            "description": "影厅ID"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "description": "开场时间"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "description": "结束时间"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "基础价格"
          },
          "available_seats": {
            "type": "integer",
            "description": "可用座位数"
          }
        }
      },
...
      "SeatMapResponse": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "integer"
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rowNo": {
                  "type": "string"
                },
                "seats": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeatStatus"
                  }
                }
              }
            }
          }
        }
      },
      "SeatStatus": {
        "type": "object",
        "properties": {
          "colNo": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "LOCKED",
              "OCCUPIED"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}